Made separate image store implementation for local folders
